# cmd/docs/swagger.yaml
basePath: /api/v1
host: localhost:8000
info:
  contact: {}
  title: Blog API
  version: "1.0"
paths:
  /user/profile:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "201":
          description: 'status": 201, "message": "Created", "data": string}'
          schema:
            type: json
        "400":
          description: 'status": 400, "message": "Bad Request", "data": null}'
          schema:
            type: json
        "409":
          description: 'status": 409, "message": "Conflict", "data": null}'
          schema:
            type: json
        "500":
          description: 'status": 500, "message": "Internal Server Error", "data":
            null}'
          schema:
            type: json
      summary: Get a profile
      tags:
      - user
  /user/search:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "201":
          description: 'status": 201, "message": "Created", "data": string}'
          schema:
            type: json
        "400":
          description: 'status": 400, "message": "Bad Request", "data": null}'
          schema:
            type: json
        "409":
          description: 'status": 409, "message": "Conflict", "data": null}'
          schema:
            type: json
        "500":
          description: 'status": 500, "message": "Internal Server Error", "data":
            null}'
          schema:
            type: json
      summary: Search users from query
      tags:
      - user
  /user/signin:
    post:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: 'status": 200, "message": "OK", "data": string}'
          schema:
            type: json
        "400":
          description: 'status": 400, "message": "Bad Request", "data": null}'
          schema:
            type: json
        "401":
          description: 'status": 401, "message": "Unauthorized", "data": null}'
          schema:
            type: json
        "404":
          description: 'status": 404, "message": "Not Found", "data": null}'
          schema:
            type: json
        "500":
          description: 'status": 500, "message": "Internal Server Error", "data":
            null}'
          schema:
            type: json
      summary: Sign in to the account
      tags:
      - user
  /user/signup:
    post:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "201":
          description: 'status": 201, "message": "Created", "data": string}'
          schema:
            type: json
        "400":
          description: 'status": 400, "message": "Bad Request", "data": null}'
          schema:
            type: json
        "409":
          description: 'status": 409, "message": "Conflict", "data": null}'
          schema:
            type: json
        "500":
          description: 'status": 500, "message": "Internal Server Error", "data":
            null}'
          schema:
            type: json
      summary: Create an account
      tags:
      - user
swagger: "2.0"